Refactor useSpeech and useForm after code review
