+++ src/components/Inputs/Buttons/VoteButton.stories.js
