+++ src/components/SignUp/SignUpForm.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import styled from 'styled-components/macro'
 import { AuthConsumer } from '../Auth/AuthContext'
-import DefaultButton from '../Inputs/DefaultButton'
+import DefaultButton from '../Inputs/Buttons/DefaultButton'
 
 export default function UserForm({ profile, setProfile, history }) {
   return (
