+++ src/components/Speech/Evaluation.js
@@ -2,7 +2,7 @@ import React from 'react'
 import styled from 'styled-components/macro'
 import useDate from '../../hooks/useDate'
 import useForm from '../../hooks/useForm'
-import DefaultButton from '../Inputs/DefaultButton'
+import DefaultButton from '../Inputs/Buttons/DefaultButton'
 import Comment from './Comment'
 import EvaluationFooter from './EvaluationFooter'
 import Statistics from './Statistics'
